Redesigned the asset management system.  The requirement of running AssetBinService at all times and the separation between src/bin data was becoming too much of a pain to deal with and complicated editor operations.  There is now only one data folder which contains the final versions of all assets.
Currently, most of the assets are just stored as json text files as they're small enough that json is sufficiently fast to load.  The exceptions are models and textures.  Model text files take way too long and textures need to be converted into DDS format, so both of these are run through the new AssetImporter.exe to convert them into better file formats.